🔧 chore(common.nix): add ranger to the list of packages for better file management 🔥 remove(mini-ai.lua): delete unused mini-ai plugin to clean up the codebase 🔨 refactor(treesitter.lua): remove unused functions and keymaps to simplify the code 👌 improve(wezterm-config.lua): decrease font size from 24.0 to 20.0 for better readability
The ranger package was added to the list of packages in common.nix to provide a more efficient way of managing files. The mini-ai plugin was removed from the codebase as it was no longer being used. In treesitter.lua, unused functions and keymaps were removed to simplify the code and make it easier to maintain. The font size in wezterm-config.lua was decreased to improve readability.