🔧 refactor(wezterm.conf.nix, common.nix): add 'Iosevka Term' font and its package 🎨 feat(dial.lua, nvim-cmp.lua, inc-rename.lua, modes.lua, nvim-spider.lua, telescope.lua, treesitter.lua, neo-tree.lua): enhance editor functionality and UI
The 'Iosevka Term' font was added to the list of fonts in wezterm.conf.nix and its package was added in common.nix for better font options.

In the editor, several new plugins and configurations were added to improve the coding experience. The dial.nvim plugin was added for better increment/decrement functionality. The nvim-cmp plugin was reconfigured for better sorting. The inc-rename.nvim and modes.nvim plugins were added for better renaming and mode handling. The nvim-spider plugin was reconfigured to skip insignificant punctuation. The keybinding for resuming Telescope results window was changed from <leader>R to <leader>T for better accessibility. The nvim-ts-autotag plugin was added for better tag handling and the nvim-biscuits plugin was added for better code navigation.

The neo-tree plugin was reconfigured to follow the current file and to bind to the current working directory for better file navigation. The visibility of filtered items was set to false for a cleaner UI.